Created a base class for the outreach bot, and found and fixed the source of the bug from the previous commit.
The base class stores variables for the hardware and has the init method that initializes the variables.  Also contains a scaleInput method.
When scaleInput takes an input, it first checks if the index it generates is less than zero, then with an else if it checks if the value is to high.
The second check should be in a separate if statement